Fix mouse rate limiting vs. potx/y polling to use time limit instead of difference limit. The original problem is better described below:
https://sourceforge.net/tracker/?func=detail&atid=1057617&aid=3590369&group_id=223021

ALeX Kazik: "I find it a bad idea to "remember" the mouse movement and
"deliver" it when asked. a real c64 does not do that. and when I copy the
current pot[xy] to the internal buffer then the remaining mouse movement is
"absorbed" (on a real c64, not vice)"


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@26590 379a1393-f5fb-40a0-bcee-ef074d9b53f7