Gtk3: model settings: implement glue logic for C64DTV widgets
The DTV model widgets now interact: setting the DTV-model will update the VIC-II and Revision widgets, while setting either the VIC-II or Revision widget will update the DTV model widget, or invalidate it in case the model is Hummmer and PAL is selected.
This involves quite some callbacks/event handlers, so the widgets should only trigger resource updates when necessary, to avoid widget A triggering events on widget B and widget A then triggering events on widget B again, and so. Hence to copious use of debug_gtk3() calls for now.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@34438 379a1393-f5fb-40a0-bcee-ef074d9b53f7