Remove use of the standard header stdbool.h
For some reason not all code linked to vice core or the UI code agrees on what a bool actually is. So let's revert to int to be on the safe side.
I haven't changed libhvsc yet, that seems to be self contained and not exporting any bool/_Bool in its API, though I'll have to check to be sure.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@35582 379a1393-f5fb-40a0-bcee-ef074d9b53f7