Gtk3: CBM-II model settings: add CRTC video standard widget
Add a 'video standard' widget for the CRTC in the CBM 6x0/7x0 models. This makes the model settings dialog in xcbm2 work a lot better, contrary to what vice.texi claims. Currently the widget is in an odd place, so the layout can definitely be improved and some more glue logic is missing. But it's a start.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@36350 379a1393-f5fb-40a0-bcee-ef074d9b53f7