Remove VICEDIR, LIBDIR, DOCDIR etc, only keep VICE_DATADIR and VICE_DOCDIR
No need for those old vars nor the new VICE_[ROOT|BIN]DIR I introduced. We have two situations: the Unix way where $prefix determines the root install dir, which can be tweaked via configure, or we have the Windows/MacOS 'stuff-all-in-a-single-dir' approach, in which the dirs are fixed relative paths.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@37437 379a1393-f5fb-40a0-bcee-ef074d9b53f7