Fix bug #1237 by disallowing to overwrite existing files.
The original problem was with unpatched PET 2001 ROMs which LOAD with SA=1,
which means to the drive that it is opening a file for WRITE. The fsdevice
happily started to overwrite existing files, which is bad.

Disadvantage: now support for SAVE-@ becomes more needed, and it seems that 
it isn't currently supported in the fsdevice.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@38018 379a1393-f5fb-40a0-bcee-ef074d9b53f7