Gtk3: Revert to normal dialog behaviour
Open means open, Autostart opens an image and then loads and start the first file. The directory preview for disks and tapes allows one to doubleclick on a file and then attach the related image and load and run the selected file. I think I've restored most of the previous.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@38631 379a1393-f5fb-40a0-bcee-ef074d9b53f7