Currently the search path includes the machine-specific roms path, the printer ROMs path, the drive ROMs path. A lot to maintain. Plus, one could theoretically move the kernal file in DRIVES and things would work.
When the common directory was introduced, files from it are loaded differently: they must be in the common/ subdirectory of archdep_get_vice_datadir().

This patch changes the loading of system files so they are loaded from a specific subdirectory of the data dir, same as what happens now for common/. So, for example, basic must be under C64 etc.

It also changes how the resource Directory (and corresponding command line option -directory) works. Before, one had to specify all of the machine-specific roms path, the printer ROMs pat and the drive ROMs path in it. Now, one just specifies the new base directory (or directories).


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@39624 379a1393-f5fb-40a0-bcee-ef074d9b53f7