Fix blocking on shutdown when running test prog: /usr/local/bin/x64sc -default -model c64c -VICIIfilter 0 -VICIIextpal -VICIIpalette pepto-pal.vpl -warp -debugcart -jamaction 1 -console -8 ../drive/rpm//rpm.d64 -pal -limitcycles 12000000 ../drive/rpm//rpm3.prg
	"  The drive has written to tracks that are not included in the currently  \n"
        "  mounted image. Do you want to write those extra tracks into the current  \n"
        "  image?"

UI thread was blocking itself from asking the user to make a choice, as the shutdown code is actually run from the original thread of the process (UI thread). Console mode now will answer UI_EXTEND_IMAGE_ALWAYS. SDL2 needs to be checked but might be ok due to not being multithreaded.

git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@40410 379a1393-f5fb-40a0-bcee-ef074d9b53f7