Gtk3: Update statusbar with second datasette widget
Incomplete due to missing UI api updates in the tapeport code, counter and state are passed only for datasette 1, so currently the widget for datasette #2 shows the counter of datasette #1. What does work is the popup menus for the widgets.
Also added a checkbox to toggle the Keyboard Status widget, there was some space due to adding the second datasette widget.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@41034 379a1393-f5fb-40a0-bcee-ef074d9b53f7