Datasette: fix static initializers
The notape_mode[2] array had in invalid initializer, only the first element was initialized to DATASETTE_CONTROL_STOP, the second was 0, which happend to work due to DATASETTE_CONTROL_STOP being defined as 0.
You cannot initialize a static array with a single non-0/NULL value, ie `static int arr[2] = { 1 }`;, only the first element will be 1, the rest will be 0.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@41050 379a1393-f5fb-40a0-bcee-ef074d9b53f7