Gtk3: Implement radio button menu items
Add support for radio buttons in the menus. Two new menu item types have been added:
- UI_MENU_TYPE_ITEM_RADIO_INT for radio groups with integer IDs
- UI_MENU_TYPE_ITEM_RADIO_STR for radio groups with string IDs

A new member has been added to ui_menu_item_t: `const char *resource`, which provides the resource name for check items and radio button items. The member `void *data` is now used for radio buttons to check against the resource to set the proper radio button in a group, the member is ignored for check buttons.
Callback functions can use `g_object_get_data(widget, "ResourceName")` to get the resource name should they require so. The action name (string) has the key "ActionName" and the 'activate' signal handler ID (gulong) has the key "HandlerID".

This commit also contains a partial implementation of the "Speed" menu items as appearing on the status bar. Since this is a larger change and branching in subversion is a shitshow I'm committing that as well.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@41590 379a1393-f5fb-40a0-bcee-ef074d9b53f7