Documentation: Remove @setfilename command from vice.texi
The @setfilename command triggers output on the terminal when processing vice.texi with texi2dvi. This is due to a bug in the texi2dvi script that uses grep to search for @setfilename but doesn't redirect grep's output to /dev/null, resulting in the @setfilename command ending up in the output of make.
The @setflename command isn't required anymore since a long time: texinfo is clever enough to take the input filename 'vice.texi' and turn it into 'vice.info'. Should generating the PDF fail on some ancient system, then this commit is where to look ;)



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@41758 379a1393-f5fb-40a0-bcee-ef074d9b53f7