nother major reworking of the CIA shift register, to make it work for the things tested by the cia-sdr-icr tests from bug #1219.
To achieve cycle-exactness of several internal processes without
creating a separate alarm callback for each (which also could have
problems with ordering them), there is now a central SDL alarm which is
called each cycle.  The same technique is used as in Wolfgang Lorenz' "A
Software Model of the CIA6526", but optimized to run only for a short
while as long as needed.

Upcoming events are scheduled by putting bits into a software version of
a mercury delay line.  The delay gets shifted every clock, and when a
bit reaches a 0-position, the task is due.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@42869 379a1393-f5fb-40a0-bcee-ef074d9b53f7