Gtk3: Refactor and improve resource browser widget, fix logfile browser
Refactor the "resource browser" code, making handling internal state easier
and more robust. Add posibility to set a default directory for the "Save"
variant of the widget: if the resource is NULL/empty the default directory
(if set) will be used as the directory for the file chooser.
This also fixes the "LogFileName" widget by setting the directory to
archdep_user_state_path() if "LogFileName" is NULL.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@42902 379a1393-f5fb-40a0-bcee-ef074d9b53f7