Gtk3: Settings dialog: pack Close button in GtkButtonBox
Packing the button inside a GtkButtonBox makes the button the same size as
actual dialog buttons added by GtkDialog in its private GtkButtonBox. We can
now also add more buttons, should we wish, and they'll appear homogeneous in
size.
Also add a mnemonic to the button, so I can close the dialog and the emu with
Alt+C,Alt+Q.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@42908 379a1393-f5fb-40a0-bcee-ef074d9b53f7