Gtk3: Settings dialog: Refactor "cartridges/ramlink"
Try to fix layout as much as possible. Still a layout issue left with the
"Browse..." buttons for the ROM and RAM images: Due to the presence of the
(slightly wider) "Save as..."/"Save image" buttons the "Browse..." button for
the ROM image is slightly less wide than the other "Browse..." button. No
amount of get_allocated_size()/get_size_allocation()/set_size)_request can fix
that.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@42985 379a1393-f5fb-40a0-bcee-ef074d9b53f7