Gtk3: Refactor string (key,value) resource combo box to use mediator
Refactor combo box for (key,value) pairs for string resource to use the resource
mediator. Greatly simplifies code and shaves off a couple of heap-allocations.
Make the string combo box use a proper model and allow instanciation with an
empty list and provide a method for the user to add values later: this will
simplify widgets that dynamically generate an initializer list on the heap and
later deallocate the list.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43066 379a1393-f5fb-40a0-bcee-ef074d9b53f7