Gtk3: Refactor hex combo box as a child of the int combo box
Since the hex combo box just has a different interface for adding values but
otherwise operates on an integer resource we can implement the hex combo box
on top of the integer combo. The hex combo box becomes a 'child' of the integer
combo box 'class', reusing a lot of code.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43073 379a1393-f5fb-40a0-bcee-ef074d9b53f7