Gtk3: Windows: Do NOT unregister CBM font with RemoveFontResourceA()
When running with Pango 1.50.2, do not call RemoveFontResourceA() on shutdown.
For some bizare reason the font remains unregistered for the remaining Windows
session and subsequent runs of VICE would require the user to reboot Windows.
I've added logging of the Pango version and the functions being called (or not
called) when registering and unregistering the CBM font.
Perhaps we can add the unregistering again when Pango 1.50.13 is released, but
I have a feeling this Windows font crap will remain broken.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43128 379a1393-f5fb-40a0-bcee-ef074d9b53f7