Gtk3: IDE64 settings: Refactor code and update layout
Split widget additions to the main grid into function adding to the layout.
Move the device-spefific widgets with their "tabs" to the bottom to make it
more clear the widgets above it are for all IDE64 devices.
Use a homogeneous column layout to make the widgets more evenly spaced and
positioned within the main grid.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43151 379a1393-f5fb-40a0-bcee-ef074d9b53f7