Gtk3: Logfile widget: use resource filechooser for filename, put "stdout" check and "open dir" buttons next to each other
Use the resource filechooser widget for the LogFileName resource. Put the
"Log to stdout" check button and the "Open directory of logfile" push button
next to each other on the same row, packing the button in a GtkButtonBox, to
get the proper amount of padding and allow for adding more buttons later, should
the need arise.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43216 379a1393-f5fb-40a0-bcee-ef074d9b53f7