Github Actions: Nuke ffmpeg in msys2 builds
It appears msys2 has dropped ffmpeg 4.x and moved to ffmpeg 5.x, which the
current VICE code cannot use. Trying to downgrade to ffmpeg 4 (and downgrade
its many dependencies) will turn into a shitshow, and updating VICE to support
ffmpeg 5.x in addition to 3.x and 4.x is not worth the trouble.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43299 379a1393-f5fb-40a0-bcee-ef074d9b53f7