Gtk3: CRT sliders: change "Reset" button into icon, reduce size when used on status bar
Also make the button always appear in the top-right cornor above the sliders,
in the settings dialog the button would be to the right of the sliders.
Keep references to all CSS providers created so the CSS doesn't have to parsed
every time the widgets are created; the providers are unref'd on emu shutdown.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43360 379a1393-f5fb-40a0-bcee-ef074d9b53f7