Optional delay to vertical retrace.
Add for non-CRTC machines (i.e. PET 2001) a compile time option to
delay the vertical blank signal by a number of cycles. About 16-24 is
probably realistic, given the HI-RES and A BRIGHT SHINING STAR demos.

By default this is disabled because it has extra overhead that almost
never is useful. Maybe we will later turn it into a runtime option or
something clever.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43618 379a1393-f5fb-40a0-bcee-ef074d9b53f7