Implement some beam-racing timing improvements for pre-CRTC PETs.
The retrace signal should now be starting and ending at the correct time,
and so should the IRQ.
Raster line data will be pre-fetched at the start of the raster line,
instead of fetched all at once at the end.
Stores to video memory that are "behind" the beam now will not be displayed yet
(with the previous method they would be displayed); stores ahead of the beam
will be.
Replaces the incorrect code from r43618.
This improves the results from Cursor's HI-RES and the "A Bright Shining Star"
demo.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43635 379a1393-f5fb-40a0-bcee-ef074d9b53f7