SDL2: Remove incorrect typedef for `SDL_JoystickID`
As it stands now with the mistyped `#ifdef USE_SDLUI2` we typedef `SDL_JoystickID`
to `int` for SDL 2.x, not 1.x, *and* the correct typedef would be `int32_t`,
not plain `int`.
Since there are other issues with SDL 1.2 and the new joystick code it's better
to remove this incorrect typedef to avoid issues in the future.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43643 379a1393-f5fb-40a0-bcee-ef074d9b53f7