Fix bug #1877.
Interrupts caused by signals from outside the CIA, such as the FLG as used by the
tape drive (but also the shift register used by fast serial) could get lost if
the "runt this once per cycle" code cia_ifr_current() had run already for that
cycle. If we detect this is the case, run some minimal duplication of the
functionality instead, to make sure the interrupt gets noticed.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43737 379a1393-f5fb-40a0-bcee-ef074d9b53f7