When disassembling, move the determination of mem_config around
so that the monitor Disassemble command uses the same mem_config as the other
monitor Memory commands (set by the monitor "bank" command). For disassembly
in cases like breakpoints, single step, CPU history and debug cpu tracing, keep
it the same (intending to show whatever the CPU is actually seeing).


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@43910 379a1393-f5fb-40a0-bcee-ef074d9b53f7