Merge branch compyx/sdl-hotkeys into trunk
This commit changes the SDL hotkeys code to use the "new" hotkeys system in
`src/arch/shared` that is already used by the Gtk3 UI. So instead of using the
"Some&path&to&a&menu&item" syntax we use UI action names like "warp-mode-toggle"
or "drive-attach-8:0". Currently the SDL .vhk files are slightly altered copies
of the Gtk3 .vhk files, with some alterations to the vhk parser (!ifdef support)
we could merge these files into sharable between all current and future UIs.
One caveat is the fact that since we now map hotkeys to UI actions, users can
no longer map hotkeys to just any menu item in SDL UI, there has to be a UI
action ID (and thus handler) associated with that item. The SDL UI currently
pops up an error message box mentioning "no UI action ID for item" when a user
attempts to map a hotkey to such an item, which is perhaps not the cleanest and
most user-friendly way of handling this. Maybe some more subtle way can be found
to inform the user which items can be associated with hotkeys and which can't.
As a consequence, making an item available for hotkeys means adding a UI action
ID/name and implementing a handler (which I'll be documenting on the Wiki).


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@44402 379a1393-f5fb-40a0-bcee-ef074d9b53f7