Add global flag `default_settings_requested`
Add global flag for code to determine if `-default` was passed on the command
line. This used to be private to the hotkeys system, but we'll also need this
flag for skipping loading of custom joymaps and system files from a user's
configuration directory.
Type is `int` for the moment like the other global flags, but should really be
`bool`. Include machine.h for the extern declaration, just like the other flags.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@44548 379a1393-f5fb-40a0-bcee-ef074d9b53f7