Gtk3: Add WiC64 settings widgets
Add widgets for WiC64 resources in the userport settings dialog. The resources
for MAC and IP don't match what the WiC64 code dumps on the terminal, and setting
the timezone in the UI doesn't appear to work since the resource setter for
"WIC64Timezone" doesn't return 0 for valid resource values.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@44573 379a1393-f5fb-40a0-bcee-ef074d9b53f7