Fix texi2html a bit.
It complained a lot about "*** unknown texi style: xref at ../../../vice/doc/html/texi2html line 1842, <FH001> line 35984" which is a really useless message because the first line number is from the script and the FH001 reference refers to the input file (which itt doesn't say) and the line number is that of the last line (probably because it reads in the whole file at once).

The complaint was because it tried to simplify style directives (looking like @style{text}) before processing @xref{text}, and was then indignant that xref isn't a style.

The solution is simple: swap the order of those two steps.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@44752 379a1393-f5fb-40a0-bcee-ef074d9b53f7