Ugly bug fix for bug #2054.
Essentially, when the screen length is set to N text lines plus 0 scan
line (CRTC_REG_VTOTALADJ == 0), the detection of getting to the end is
done 1 scan line too late. For other values of VTOTALADJ the detection
is ok.  Just the case of 0 is missed the first time around by an
apparently unfortunate order of cheking conditions.

The plan is to clean up the hack by rearranging the order of the code
into something nice and without goto.



git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@45219 379a1393-f5fb-40a0-bcee-ef074d9b53f7