Gtk3: make the REU enable button update save-as/flush buttons' state
The cart-can-save/flush functions for REU only return `true` when the REU
emulation is enabled, so initially selecting a REU image and then enabling
REU emulation didn't enable the save-as/flush buttons (see bug #2070).
This commit should fix that. If we find other cases like this we might consider
writing a helper function to set up an extra event handler on an "Enable"
button that then calls `cart_image_widget_update_sensitivity()`.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@45337 379a1393-f5fb-40a0-bcee-ef074d9b53f7