Limit number of icon files installed
Instead of installing all icon files in data/common/, install only those
required, either by VICE itself or the OS'es guidelines.
For SDL the 256x256 PNG icons are installed, except on Windows for SDL 1.x,
where the 32x32 icons are installed (as suggested by the SDL docs).
For Gtk3 on Unix the 32x32, 48x48, 64x64 and 256x256 icons are installed,
either in their correct place when configured with `--enable-desktop-files`
or in $PREFIX/share/vice otherwise.
For Gtk3 on Windows the suggested icon sizes are used (16x16, 24x24, 32x32,
48x48, 64x64 and 256x256), linked with the executable using `icotool` and
`windres`.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@45499 379a1393-f5fb-40a0-bcee-ef074d9b53f7