Gtk3: resource spin button: connect "input" and "output" signals locked
The "input" and "output" signals of a spin button deal with parsing and formatting
of spin button text (the internal GtkEntry); to speed things up in the UI these
were connected "unlocked". However, it could be the GtkEntry updates trigger
a "value-changed" event, which in turn would alter the bound resource of the
widget, which could trigger an assertion of not having the main lock.
So to be safe, we connect these "locked". See bug #2144.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@45697 379a1393-f5fb-40a0-bcee-ef074d9b53f7