monitor: fix bug 2177
If we are leaving the monitor with a change in program flow, and there is a breakpoint on the new address, take the breakpoint right away instead of leaving the monitor.
The monitor command "G XXXX" is considered a change in flow.
This is because of the structure of the DO_INTERRUPT() macro, which can enter the monitor under different conditions in different places, but each only once per executed instruction.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk/vice@45822 379a1393-f5fb-40a0-bcee-ef074d9b53f7